# Api/messages_controller.go
package Api

import (
	"github.com/gin-gonic/gin"
	"net/http"
	"quay-go-api/Entities/Dto"
	"quay-go-api/Services"
	"quay-go-api/Services/Auth"
)

func messagesController() {
	messages := engine.Group("/api/v1/messages")
	{
		messages.Use(authorizedMiddleware)
		messages.GET("", listMessages)
		messages.POST("", createMessage)
		messages.PATCH("/:uuid", updateMessage)
		messages.DELETE("/:uuid", deleteMessage)
	}
}

// listMessages List messages displayed on the quay web app for all user
// @Description List messages displayed on the quay web app for all user
// @Summary List messages displayed on the quay web app for all user
// @Tags Messages
// @Param severity query string false "Filter message by severity (comma separated, accepted values: info, warning, error)"
// @Success 200 {object} []Dto.Message
// @Failure 401 {object} Errors.ErrorResponse "Unauthorized"
// @Failure 500 {object} Errors.ErrorResponse "Internal Server Error"
// @Security ApiKeyAuth
// @Router /api/v1/messages [get]
func listMessages(c *gin.Context) {
	_, hasScopeErr := retrieveCurrentUser(c, []Auth.Scope{Auth.OrgAdmin})
	if hasScopeErr != nil {
		throwError(c, hasScopeErr)
		return
	}

	// Get filters from query params
	filters := extractFilters(c)

	messages, err := Services.ListMessages(filters)
	if err != nil {
		throwError(c, err)
		return
	}

	c.JSON(http.StatusOK, messages)
}

// createMessage Create a new message to displayed on the quay web app for all user
// @Description Create a new message to displayed on the quay web app for all user
// @Summary Create a new message to displayed on the quay web app for all user
// @Tags Messages
// @Accept json
// @Param message body Dto.CreateMessage true "Message content and severity"
// @Success 201 {object} Dto.Message
// @Failure 401 {object} Errors.ErrorResponse "Unauthorized"
// @Failure 500 {object} Errors.ErrorResponse "Internal Server Error"
// @Security ApiKeyAuth
// @Router /api/v1/messages [post]
func createMessage(c *gin.Context) {
	hasScopesErr := requiredScopes(c, []Auth.Scope{Auth.AdminUser})
	if hasScopesErr != nil {
		throwError(c, hasScopesErr)
		return
	}

	var createMessage Dto.CreateMessage
	_ = c.BindJSON(&createMessage)

	createdMessage, err := Services.CreateMessage(createMessage)
	if err != nil {
		throwError(c, err)
		return
	}
	c.JSON(http.StatusCreated, createdMessage)
}

// updateMessage Update the content or the severity of an existing message displayed on the quay web app for all user
// @Description Update the content or the severity of an existing message displayed on the quay web app for all user
// @Summary Update the content or the severity of an existing message displayed on the quay web app for all user
// @Tags Messages
// @Accept json
// @Param uuid path string true "UUID of the message to update"
// @Param update body Dto.UpdateMessage true "Message content and severity"
// @Success 200 {object} Dto.Message
// @Failure 401 {object} Errors.ErrorResponse "Unauthorized"
// @Failure 500 {object} Errors.ErrorResponse "Internal Server Error"
// @Security ApiKeyAuth
// @Router /api/v1/messages/{uuid} [patch]
func updateMessage(c *gin.Context) {
	hasScopesErr := requiredScopes(c, []Auth.Scope{Auth.AdminUser})
	if hasScopesErr != nil {
		throwError(c, hasScopesErr)
		return
	}

	messageUUID := c.Param("uuid")

	var updateMessage Dto.UpdateMessage
	_ = c.BindJSON(&updateMessage)

	updatedMessage, err := Services.UpdateMessage(messageUUID, updateMessage)
	if err != nil {
		throwError(c, err)
		return
	}
	c.JSON(http.StatusOK, updatedMessage)
}

// deleteMessage Delete a message
// @Description Delete a message
// @Summary Delete a message
// @Tags Messages
// @Param uuid path string true "UUID of the message to update"
// @Success 204 "No Content"
// @Failure 401 {object} Errors.ErrorResponse "Unauthorized"
// @Failure 500 {object} Errors.ErrorResponse "Internal Server Error"
// @Security ApiKeyAuth
// @Router /api/v1/messages/{uuid} [delete]
func deleteMessage(c *gin.Context) {
	hasScopesErr := requiredScopes(c, []Auth.Scope{Auth.AdminUser})
	if hasScopesErr != nil {
		throwError(c, hasScopesErr)
		return
	}

	messageUUID := c.Param("uuid")

	err := Services.DeleteMessage(messageUUID)
	if err != nil {
		throwError(c, err)
		return
	}
	c.Status(http.StatusNoContent)
}