# docs/swagger.yaml
definitions:
  Dto.Avatar:
    properties:
      color:
        description: hex color code
        type: string
      hash:
        type: string
      kind:
        description: e.g., "user", "org"
        type: string
      name:
        type: string
    type: object
  Dto.CreateMessage:
    properties:
      content:
        type: string
      severity:
        type: string
    type: object
  Dto.CreateOrganization:
    properties:
      name:
        type: string
    type: object
...
  Dto.Message:
    properties:
      content:
        type: string
      media_type:
        type: string
      severity:
        type: string
      uuid:
        type: string
    type: object
  Dto.Organization:
    properties:
      avatar:
        $ref: '#/definitions/Dto.Avatar'
      invoice_email:
        type: boolean
      invoice_email_address:
        type: string
      is_admin:
        type: boolean
      is_free_account:
        type: boolean
      is_member:
        type: boolean
      name:
        type: string
      tag_expiration_s:
        type: integer
      teams:
        items:
          $ref: '#/definitions/Dto.Team'
        type: array
    type: object
  Dto.OrganizationMember:
    properties:
      avatar:
        $ref: '#/definitions/Dto.Avatar'
      kind:
        type: string
      name:
        type: string
      repositories:
        description: List of member's repository name
        items:
          type: string
        type: array
      teams:
        description: List of member's team name
        items:
          type: string
        type: array
    type: object
...
  Dto.Team:
    properties:
      avatar:
        $ref: '#/definitions/Dto.Avatar'
      can_view:
        type: boolean
      description:
        type: string
      is_synced:
        type: boolean
      members_count:
        description: RepoCount    int    `json:"repo_count"` // Useless?
        type: integer
      name:
        type: string
      role:
        type: string
    type: object
...
  Dto.UpdateMessage:
    properties:
      content:
        type: string
      severity:
        type: string
    type: object
  Dto.UpdateOrganization:
    properties:
      email:
        type: string
      invoice_email:
        type: boolean
      invoice_email_address:
        type: string
      tag_expiration_s:
        type: integer
    type: object
...
  Dto.UserOrganization:
    properties:
      avatar:
        $ref: '#/definitions/Dto.Avatar'
      can_create_repo:
        type: boolean
      is_org_admin:
        type: boolean
      name:
        type: string
      preferred_namespace:
        type: boolean
      public:
        type: boolean
    type: object
  Errors.ErrorDetails:
    properties:
      code:
        description: Code is a sting that represents the error code (e.g. "USER_NOT_FOUND",
          "INVALID_INPUT", etc.)
        type: string
      message:
        type: string
    type: object
  Errors.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/Errors.ErrorDetails'
    type: object
info:
  contact: {}
  description: Quay registry API implemented in Go
  title: Quay Go API
  version: "1.0"
paths:
  /api/v1/messages:
    get:
      description: List messages displayed on the quay web app for all user
      parameters:
      - description: 'Filter message by severity (comma separated, accepted values:
          info, warning, error)'
        in: query
        name: severity
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Dto.Message'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List messages displayed on the quay web app for all user
      tags:
      - Messages
    post:
      consumes:
      - application/json
      description: Create a new message to displayed on the quay web app for all user
      parameters:
      - description: Message content and severity
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/Dto.CreateMessage'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Dto.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new message to displayed on the quay web app for all user
      tags:
      - Messages
  /api/v1/messages/{uuid}:
    delete:
      description: Delete a message
      parameters:
      - description: UUID of the message to update
        in: path
        name: uuid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a message
      tags:
      - Messages
    patch:
      consumes:
      - application/json
      description: Update the content or the severity of an existing message displayed
        on the quay web app for all user
      parameters:
      - description: UUID of the message to update
        in: path
        name: uuid
        required: true
        type: string
      - description: Message content and severity
        in: body
        name: update
        required: true
        schema:
          $ref: '#/definitions/Dto.UpdateMessage'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Dto.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update the content or the severity of an existing message displayed
        on the quay web app for all user
      tags:
      - Messages
  /api/v1/organization:
    get:
      description: List user's organizations
      parameters:
      - description: Filter organizations by public or private (true/false)
        in: query
        name: is_public
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Dto.UserOrganization'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List user's organizations
      tags:
      - Organization
    post:
      consumes:
      - application/json
      description: Create a new organization
      parameters:
      - description: Organization metadata
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/Dto.CreateOrganization'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Dto.Organization'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new organization
      tags:
      - Organization
  /api/v1/organization/{orgname}:
    delete:
      description: Delete an organization
      parameters:
      - description: Name of the organization
        in: path
        name: orgname
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete an organization
      tags:
      - Organization
    get:
      description: Get details in an organization
      parameters:
      - description: Name of the organization
        in: path
        name: orgname
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Dto.Organization'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get details in an organization
      tags:
      - Organization
    patch:
      consumes:
      - application/json
      description: Update organization details
      parameters:
      - description: Name of the organization
        in: path
        name: orgname
        required: true
        type: string
      - description: Organization details to change
        in: body
        name: update
        required: true
        schema:
          $ref: '#/definitions/Dto.UpdateOrganization'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Dto.Organization'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update organization details
      tags:
      - Organization
  /api/v1/organization/{orgname}/members:
    get:
      description: List organization's members
      parameters:
      - description: Name of the organization
        in: path
        name: orgname
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Dto.OrganizationMember'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List organization's members
      tags:
      - Organization
...
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication using the Authorization header. The value
      should be in the format "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"